Move self-signed cert fixture to utils.py (#814)
