--- tests/load/filesystem/conftest.py
@@ -1,49 +1,0 @@
-import os
-import tempfile
-from typing import Iterator
-
-import pytest
-from cryptography import x509
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.x509.oid import NameOID
-
-from dlt.common.pendulum import timedelta, __utcnow
-
-
-@pytest.fixture(scope="function")
-def self_signed_cert() -> Iterator[str]:
-    key = rsa.generate_private_key(public_exponent=65537, key_size=2048, backend=default_backend())
-
-    subject = issuer = x509.Name(
-        [
-            x509.NameAttribute(NameOID.COUNTRY_NAME, "DE"),
-            x509.NameAttribute(NameOID.STATE_OR_PROVINCE_NAME, "Berlin"),
-            x509.NameAttribute(NameOID.LOCALITY_NAME, "Britz"),
-            x509.NameAttribute(NameOID.ORGANIZATION_NAME, "dltHub"),
-            x509.NameAttribute(NameOID.COMMON_NAME, "localhost"),
-        ]
-    )
-    cert = (
-        x509.CertificateBuilder()
-        .subject_name(subject)
-        .issuer_name(issuer)
-        .public_key(key.public_key())
-        .serial_number(x509.random_serial_number())
-        .not_valid_before(__utcnow())
-        .not_valid_after(__utcnow() + timedelta(days=1))
-        .add_extension(
-            x509.SubjectAlternativeName([x509.DNSName("localhost")]),
-            critical=False,
-        )
-        .sign(key, hashes.SHA256(), default_backend())
-    )
-
-    with tempfile.NamedTemporaryFile(suffix=".crt", delete=False) as cert_file:
-        cert_file.write(cert.public_bytes(serialization.Encoding.PEM))
-        cert_path = cert_file.name
-
-    yield cert_path
-
-    os.remove(cert_path)
